updates to auton turning function

+++ include/subsystemHeaders/drive.hpp
@@ -10,4 +10,4 @@ void motorDrive();
 void moveForward(double units, int voltage);
 void moveBack(double units, int voltage);
 void resetEncoders();
-void turn(double degrees, int voltage);
+void turn(int degrees, int voltage);
